MCP/code/MCP-DOCS-AGENT/requirements.txt
mcp
httpx
python-dotenv
beautifulsoup4
langchain
langgraph
langchain-ollama
langchain-mcp-adapters
